rubocop warnings solved, partial reek warnings fixed
