+++ app/domain/mappers/airport_mapper.rb
@@ -54,4 +54,3 @@ module ComfyWings
     end
   end
 end
-
