--- app/presentation/view_objects/flights.rb
@@ -1,22 +1,0 @@
-# frozen_string_literal: true
-
-module Views
-  # class for trip info
-  class Trips
-    def initialize(trip)
-      @trip = trip
-    end
-
-    def origin
-      @trip.origin
-    end
-
-    def destination
-      @flight.destination
-    end
-
-    def flights
-      Flight.new(@trip.flights)
-    end
-  end
-end
