+++ config/environment.rb
@@ -32,6 +32,7 @@ module ComfyWings
 
       # Database Setup
       DB = Sequel.connect(ENV.fetch('DATABASE_URL'))
+      # deliberately :reek:UncommunicativeMethodName calling method DB
       def self.DB = DB # rubocop:disable Naming/MethodName
 
       # Setup for logger
