+++ spec/helpers/database_helper.rb
@@ -2,6 +2,7 @@
 
 # Helper to clean database during test runs
 module DatabaseHelper
+  # Deliberately :reek:DuplicateMethodCall on App.DB
   def self.wipe_database
     # Ignore foreign key constraints when wiping tables
     ComfyWings::App.DB.run('PRAGMA foreign_keys = OFF')
